MOB-1165 (Mutual auth commands do not complete) - added back SecureCommandTEst and a gigpaxos jar with the mutual auth fix
